Move MockQuietWhiteboard into mocks.go
